Enum thrift_codec::data::Elements

pub enum Elements {
    Bool(Vec<bool>),
    I8(Vec<i8>),
    I16(Vec<i16>),
    I32(Vec<i32>),
    I64(Vec<i64>),
    Double(Vec<f64>),
    Binary(Vec<Vec<u8>>),
    Struct(Vec<Struct>),
    Map(Vec<Map>),
    Set(Vec<Set>),
    List(Vec<List>),
}

A sequence of the values of a data kind.

Variants

Bool(Vec<bool>)

I8(Vec<i8>)

I16(Vec<i16>)

I32(Vec<i32>)

I64(Vec<i64>)

Double(Vec<f64>)

Binary(Vec<Vec<u8>>)

Struct(Vec<Struct>)

Map(Vec<Map>)

Set(Vec<Set>)

List(Vec<List>)

Methods

impl Elements

pub fn new(kind: DataKind) -> Self

Makes an empty sequence which can have the elements belonging to kind.

pub fn get(&self, index: usize) -> Option<DataRef>

Returns the element placed at the specified index.

pub fn len(&self) -> usize

Returns the element count of this sequence.

pub fn is_empty(&self) -> bool

Returns true if this sequence has no elements.

Important traits for ElementIter<'a>
impl<'a> Iterator for ElementIter<'a> type Item = DataRef<'a>;

pub fn iter(&self) -> ElementIter

Returns an iterator over the elements of this sequence.

pub fn kind(&self) -> DataKind

Returns the kind of the elements in this sequence.