Struct thrift_codec::message::Message

pub struct Message { /* fields omitted */ }

RPC message.

Methods

impl Message

pub fn new(
method_name: &str,
kind: MessageKind,
sequence_id: i32,
body: Struct
) -> Self

Makes a new Message instance.

pub fn call(method_name: &str, sequence_id: i32, body: Struct) -> Self

Makes a new Message instance which has the kind MessageKind::Call.

pub fn reply(method_name: &str, sequence_id: i32, body: Struct) -> Self

Makes a new Message instance which has the kind MessageKind::Reply.

pub fn exception(method_name: &str, sequence_id: i32, body: Struct) -> Self

Makes a new Message instance which has the kind MessageKind::Exception.

pub fn oneway(method_name: &str, sequence_id: i32, body: Struct) -> Self

Makes a new Message instance which has the kind MessageKind::Oneway.

pub fn method_name(&self) -> &str

Returns the method name of this message.

pub fn kind(&self) -> MessageKind

Returns the kind of this message.

pub fn sequence_id(&self) -> i32

Returns the sequence id of this message.

pub fn body(&self) -> &Struct

Returns the body of this message.